requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==46.0.1
discord.py==2.3.2
dj-database-url==2.1.0
Django==5.2.6
firebase-admin==6.2.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
msgpack==1.1.1
multidict==6.6.4
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0
yarl==1.20.1
